Updated error handling and added documentation.
