--- LTU.SearchEngine.Frontend/src/classes/customeError.ts
@@ -1,9 +1,0 @@
-export class CustomError extends Error {
-    errorCode: number;
-
-    constructor( errorCode: number, message: string ){
-        super(message);
-        this.errorCode = errorCode;
-        this.name = 'CustomError';
-    }
-}
